Don't allow silent errors in xfs_inactive().
xfs_inactive() fails to report errors when committing the inactive
transaction. Hence we can get silent failures either finishing off
the truncation or committing the transaction. Even if we get errors,
we need to continue, so simply warn loudly to the system if we get
errors here.
Merge of xfs-linux-melb:xfs-kern:30830a by kenmcd.

  Don't allow silent errors in xfs_inactive().