Sanitise xfs_log_force error checking.
xfs_log_force() is declared to return an error, but we almost never
check it. We don't need to check it in most cases; if there's a log
I/O error then we'll be shutting down the filesystem anyway and that
means we'll catch the error somewhere else.

However, on certain calls we should be returning an error - sync
transactions, fsync, sync writes, etc. so this isn't a pure black
and white distinction. Hence make xfs_log_force() a void function
that issues a warning to the syslog on error, and call
_xfs_log_force() in all the places where we actually care about the
error status returned.
Merge of xfs-linux-melb:xfs-kern:30832a by kenmcd.

  Sanitise xfs_log_force error checking.