Always use di_forkoff when checking for attr space.
In the case where we mount a filesystem which was previously
using the attr2 format as attr1, returning the default
mp->m_attroffset instead of the per-inode di_forkoff for
inline attribute fit calculations, may result in corruption,
if for example, the data fork is already taking more space
than the default fork offset and we try to add an extended attribute.
Fix tested by xfstests/186.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>
Merge of xfs-linux-melb:xfs-kern:30861a by kenmcd.

  Fix up xfs_attr_shortform_bytesfit() to return di_forkoff for attr1 instead
  of m_attroffset, when we have a fit in the inode.
  Fix tested by xfstests/186.