xfs_bmap_compute_maxlevels should be based on di_forkoff instead of m_attr_offset
Fix up xfs_bmap_compute_maxlevels() to account for the case when we go
from using attr2 to using attr1. In that case attr1 will no longer necessarily
be at m_attr_offset>>3, but could be at a different value for di_forkoff.
Therefore, we return the worst case scenario using MINDBTPTRS and MINABTPTRS,
as this function is used for determining the maximum log space.
Merge of xfs-linux-melb:xfs-kern:30862a by kenmcd.

  Fix up xfs_bmap_compute_maxlevels() to account for the case when we go
  from using attr2 to using attr1. In that case attr1 will no longer necessarily
  be at m_attr_offset>>3, but could be at a different value for di_forkoff.
  Therefore, we return the worst case scenario using MINDBTPTRS and MINABTPTRS,
  as this function is used for determining the maximum log space.