Remove CONFIG_XFS_SECURITY.
There is no point to the CONFIG_XFS_SECURITY option; it
disables the ability to set security attributes at runtime,
but it does not actually slim down or remove any code for
runtime.  Just remove it and always allow security attributes
to be set.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>
Merge of xfs-linux-melb:xfs-kern:30877a by kenmcd.

  remove CONFIG_XFS_SECURITY