remove manual lookup from xfs_rename and simplify locking
->rename already gets the target inode passed if it exits.  Pass it down
to xfs_rename so that we can avoid looking it up again.  Also simplify
locking as the first lock section in xfs_rename can go away now:  the
isdir is an invariant over the lifetime of the inode, and new_parent and
the nlink check are namespace topology protected by i_mutex in the VFS.
The projid check needs to move into the second lock section anyway to
not be racy.

Also kill the now unused xfs_dir_lookup_int and remove the now-unused
first_locked argumet to xfs_lock_inodes.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:30903a by kenmcd.

  remove manual lookup from xfs_rename and simplify locking