kill usesless IHOLD calls in xfs_rename
Similar to to the previous patch for remove and rmdir only grab
a reference to inodes when we join them to transaction to balance the
decrement on transaction completion.  Everything else it taken care of
by the VFS.

Note that the old case had leaks of inode count when src == target or
src or target == one of the parent inodes, but these cases are
fortunately already rejected by the VFS.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:30904a by kenmcd.

  kill usesless IHOLD calls in xfs_rename