kill usesless IHOLD calls in xfs_remove and xfs_rmdir
The VFS always has an inode reference when we call these functions.  So
we only need to grab a signle reference to each inode that's joined to
a transaction - all the other bumping and dropping is as useless as the
comments describing the IRIX semantics.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:30912a by kenmcd.

  kill usesless IHOLD calls in xfs_remove and xfs_rmdir