Don't double count reserved block changes on UP.
On uniprocessor machines, the incore superblock is used for
all in memory accounting of free blocks. in this situation,
changes to the reserved block count are accounted twice;
once directly and once via xfs_mod_incore_sb(). Seeing as
the modification on SMP is done via xfs_mod_incore_sb(),
make this the only update mechanism that UP uses as well.
Merge of xfs-linux-melb:xfs-kern:30997a by kenmcd.

  Don't double count reserved block changes on UP.