Don't initialise new inode generation numbers to zero
When we allocation new inode chunks, we initialise the generation
numbers to zero. This works fine until we delete a chunk and then
reallocate it, resulting in the same inode numbers but with a
reset generation count. This can result in inode/generation
pairs of different inodes occurring relatively close together.

Given that the inode/gen pair makes up the "unique" portion of
an NFS filehandle on XFS, this can result in file handles cached
on clients being seen on the wire from the server but refer to
a different file. This causes .... issues for NFS clients.

Hence we need a unique generation number initialisation for
each inode to prevent reuse of a small portion of the generation
number space. Use a random number to initialise the generation
number so we don't need to keep any new state on disk whilst
making the new number difficult to guess from previous allocations.
Merge of xfs-linux-melb:xfs-kern:31001a by kenmcd.

  use random32() to initialise the generation in newly allocated
  inodes to prevent short term reuse of inode,gen pairs which
  can cause ESTALE problems for NFS clients.