Fix up noattr2 so that it will properly update the versionnum and features2 fields.
Previously, mounting with noattr2 failed to achieve anything because
although it cleared the attr2 mount flag, it would set it again as soon as
it processed the superblock fields.
The fix now has an explicit noattr2 flag and uses it later to fix up the
versionnum and features2 fields.
Merge of xfs-linux-melb:xfs-kern:31003a by kenmcd.

  Add in flag XFSMNT_NOATTR2.