xfs_dm_rdwr() needs to pass a vfsmount to dentry_open()
We need a vfsmount pointer in xfs_dm_rdwr() but we are not
provided with one and there's no way to get to it.  So add
a m_vfsmount field to the xfs_mount structure and set it
up at mount time.  We can then access it through the xfs
inode.
Merge of xfs-linux-melb:xfs-kern:31176a by kenmcd.

  Set up a vfsmount pointer in the xfs mount structure.