make inode reclaim wait for log I/O to complete
During a forced shutdown a xfs inode can be destroyed before log
I/O involving that inode is complete.  We need to wait for the
inode to be unpinned before tearing it down.  Version 2 cleans up
the code a bit by relying on xfs_iflush() to do the unpinning and
forced shutdown check.
Merge of xfs-linux-melb:xfs-kern:31326a by kenmcd.

  make inode reclaim wait for log I/O to complete