Always reset btree cursor after an insert
After a btree insert operation a cursor can be invalid due to block
splits and a maybe a new root block.  We reset the cursor in
xfs_bmbt_insert() in the cases where we think we need to but it
isn't enough as we still see assertions.  Just do what we do elsewhere
and reset the cursor unconditionally.  Also remove the fix to
revalidate the original cursor in xfs_bmbt_insert().
Merge of xfs-linux-melb:xfs-kern:31342a by kenmcd.

  Always reset btree cursor after an insert