attrmulti cleanup
xfs_attrmulti_by_handle currently request the size based on
sizeof(attr_multiop_t) but should be using sizeof(xfs_attr_multiop_t)
because that is what it is dealing with.  Despite beeing wrong this
actually harmless in practice because both structures are the same size
on all platforms.

But this sizeof was the only user of struct attr_multiop so we can just
kill it.  Also move the ATTR_OP_* defines xfs_attr.h into the
struct xfs_attr_multiop defintion in xfs_fs.h because they are only used
with that structure, and are part of the user ABI for the
XFS_IOC_ATTRMULTI_BY_HANDLE ioctl.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:31352a by kenmcd.

  attrmulti cleanup