Restore the lowspace extent allocator algorithm
When free space is running low the extent allocator may choose to
allocate an extent from an AG without leaving sufficient space for
a btree split when inserting the new extent (see where
xfs_bmap_btalloc() sets minleft to 0).  In this case the allocator
will enable the lowspace algorithm which is supposed to allow further
allocations (such as btree splits and newroots) to allocate from
sequential AGs.  This algorithm has been broken for a long time
and this patch restores its behaviour.
Merge of xfs-linux-melb:xfs-kern:31358a by kenmcd.

  Restore the lowspace extent allocator algorithm