Do not access buffers after dropping reference count
We should not access a buffer after dropping it's reference count otherwise
we could race with another thread that releases the final reference count
and frees the buffer causing us to access potentially unmapped memory.
The bug this change fixes only occured on DEBUG XFS since the offending
code was in an ASSERT.
Merge of xfs-linux-melb:xfs-kern:31715a by kenmcd.

  Do not access buffers after dropping reference count