move root inode IRELE into xfs_unmountfs
The root inode is allocated in xfs_mountfs so it should be release in
xfs_unmountfs.  For the unmount case that means we do it after the
the xfs_sync(mp, SYNC_WAIT | SYNC_CLOSE) in the forced shutdown case
and the dmapi unmount event.  Note that both reference the rip variable
which might be freed by that time in case inode flushing has kicked in,
so strictly speaking this might count as a bug fix


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:31830a by kenmcd.

  move root inode IRELE into xfs_unmountfs