refactor xfs_mount_free
xfs_mount_free mostly frees the perag data, which is something that
is duplicated in the mount error path.

Move the XFS_QM_DONE call to the caller and remove the useless
mutex_destroy/spinlock_destroy calls so that we can re-use it for
the mount error path.  Also rename it to xfs_free_perag to reflect what
it does.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:31836a by kenmcd.

  refactor xfs_mount_free