Use KM_NOFS for debug trace buffers
Use KM_NOFS to prevent recursion back into the filesystem which can
cause deadlocks.

In the case of xfs_iread() we hold the lock on the inode cluster buffer
while allocating memory for the trace buffers.  If we recurse back into
XFS to flush data that may require a transaction to allocate extents
which needs log space.  This can deadlock with the xfsaild thread which
can't push the tail of the log because it is trying to get the inode
cluster buffer lock.
Merge of xfs-linux-melb:xfs-kern:31838a by kenmcd.

  Use KM_NOFS for debug trace buffers