implement generic xfs_btree_increment
From: Dave Chinner <dgc@sgi.com>

Because this is the first major generic btree routine this patch
includes some infrastrucure, first a few routines to deal with
a btree block that can be either in short or long form, second
xfs_btree_read_buf_block, which is the new central routine to read
a btree block given a cursor, and third the new xfs_btree_ptr_addr
routine to calculate the address for a given btree pointer record.

[hch: split out from bigger patch and minor adaptions]

Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:32190a by kenmcd.

  implement generic xfs_btree_increment