move xfs_bmbt_newroot to common code
xfs_bmbt_newroot is a mostly generic implementation of moving from
an inode root to a real block based root.  So move it to xfs_btree.c
where it can use all the nice infrastructure there and make it pointer
size agnostic

The new name for it is xfs_btree_new_iroot, following the old naming
but making it clear we're dealing with the root in inode case here,
and to avoid confusion with xfs_btree_new_root which is used for the
not inode rooted case.


Signed-off-by: Christoph Hellwig <hch@lst.de>
Merge of xfs-linux-melb:xfs-kern:32201a by kenmcd.

  move xfs_bmbt_newroot to common code